digital_breakthrough_sochi_2024/backend/requirements.txt
aiohttp==3.9.4
aiohttp-apispec==2.2.3
aiohttp-session==2.12.0
alembic==1.13.1
aspose-words==24.4.0
asyncpg==0.29.0
cryptography==42.0.5
marshmallow==3.21.1
pymystem3==0.2.0
PyYAML==6.0.1
SQLAlchemy==2.0.29